\name{PanelQGARCHwithFactor-Data}
\alias{Y_N_T}
\alias{Factors_r_T}
\alias{Initial_tau1per_N_8}
\alias{Initial_tau2.5per_N_8}
\alias{Initial_tau97.5per_N_8}
\alias{Initial_tau99per_N_8}
\title{Data set}
\description{
  "Y_N_T" is the dataset of 30 industry portfolios from January 3, 2012 to June 28, 2024, and is downloaded from Kenneth French’s database.
  "Factors_r_T" is the dataset of the Fama-French five factors (Fama and French, 2015).
  "Initial_tau1per_N_8", "Initial_tau2.5per_N_8", "Initial_tau97.5per_N_8" and "Initial_tau99per_N_8" are the initial estimates in paper at the lower and upper 1\% and 2.5\% quantile levels, respectively. 
}
\author{Wenyu Li}

