\name{PanelQGARCHwithFactor-package}
\alias{PanelQGARCHwithFactor-package}
\alias{PanelQGARCHwithFactor}
\docType{package}
\title{\packageTitle{PanelQGARCHwithFactor}}
\description{\packageDescription{PanelQGARCHwithFactor}}
\section{Package Content}{\packageIndices{PanelQGARCHwithFactor}}
\author{\packageAuthor{PanelQGARCHwithFactor}}
\section{Maintainer}{\packageMaintainer{PanelQGARCHwithFactor}}
\keyword{package}
